// api/src/common/enums/AtBat.ts
export enum AtBat {
   double = 12,
   doublePlay = 1,
   flyball = 2,
   foulPopfly = 3,
   groundRuleDouble = 13,
   groundout = 4,
   hitByPitch = 9,
   homeRun = 15,
   lineout = 6,
   popfly = 5,
   single = 10,
   strikeout = 7,
   triple = 14,
   triplePlay = 8,
   unknown = 16,
   walk = 11,
}