// api/src/common/enums/Pitch.ts
export enum Pitch {
   automaticIntentionalBall = 16,
   ball = 4,
   ballPutIntoPlayByBatter = 5,
   ballPutIntoPlayOnPitchout = 17,
   calledStrike = 1,
   foul = 3,
   foulBallOnPitchout = 14,
   foulBunt = 10,
   foulTip = 6,
   foulTipOnBunt = 12,
   hitBatter = 9,
   intentionalBall = 8,
   missedBuntAttempt = 11,
   swingingOnPitchout = 13,
   swingingStrike = 2,
   unknownOrMissedPitch = 15,
   unknownStrike = 7,
}